delete local account remove its files

+++ core/utils.py
@@ -1,6 +1,8 @@
 import random
 import string
 import boto3
+import shutil
+import os
 
 from evsapp import settings
 
@@ -13,7 +15,7 @@ def generate_random_string(chars=ALPHANUMERIC_CHARS, length=STRING_LENGTH):
 
 
 def delete_related_user_files_s3(username):
-    user_objects_prefix = settings.AWS_PUBLIC_MEDIA_LOCATION + '/' + username
+    user_objects_prefix = settings.AWS_PUBLIC_MEDIA_LOCATION + f'/{username}'
     bucket_name = settings.AWS_STORAGE_BUCKET_NAME
     s3 = boto3.resource('s3')
     bucket = s3.Bucket(bucket_name)
@@ -21,8 +23,14 @@ def delete_related_user_files_s3(username):
     objects_to_delete = []
     for file in s3.Bucket(settings.AWS_STORAGE_BUCKET_NAME).objects.filter(Prefix=user_objects_prefix):
         objects_to_delete.append({'Key': file.key})
+    if objects_to_delete:
+        bucket.delete_objects(Delete={
+            'Objects': objects_to_delete
+        }
+        )
 
-    bucket.delete_objects(Delete={
-        'Objects': objects_to_delete
-    }
-    )
+
+def delete_related_user_files_local(username):
+    user_dir_path = settings.MEDIA_ROOT + f'/{username}'
+    if os.path.exists(user_dir_path):
+        shutil.rmtree(user_dir_path)
